requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
APScheduler==3.11.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.0
Flask-PyMongo==3.0.1
Flask-SocketIO==5.5.1
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
mailgun==1.0.0
MarkupSafe==3.0.2
multidict==6.4.3
packaging==25.0
pika==1.3.2
propcache==0.3.1
PyJWT==2.10.1
pymongo==4.13.0
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
requests==2.32.3
simple-websocket==1.1.0
twilio==9.6.1
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
